+++ models/src/v2_0_0/tooltip/obtained_item.rs
@@ -5,12 +5,12 @@ use super::Condition;
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ObtainedEffectItem {
     pub name: String,
-    pub count: u64,
+    pub count: u32,
     pub condition: Option<Condition>,
 }
 
 impl ObtainedEffectItem {
-    pub fn new<T: Into<String>>(name: T, count: u64) -> Self {
+    pub fn new<T: Into<String>>(name: T, count: u32) -> Self {
         Self {
             name: name.into().to_title_case(),
             count,
@@ -18,7 +18,7 @@ impl ObtainedEffectItem {
         }
     }
 
-    pub fn new_conditional<T: Into<String>>(name: T, count: u64, condition: Condition) -> Self {
+    pub fn new_conditional<T: Into<String>>(name: T, count: u32, condition: Condition) -> Self {
         Self {
             name: name.into().to_title_case(),
             count,
